updated chapter 2 main.dialogue for a Trolum
John Blue updating this:

Original text: Well, Travin just came by saying that they found this weird Geode while scavenging. It was a square rock with these weird items on the right and a transparent rock on the left.

Updated text: Well, Travin just came by saying that they found some weird Geode while scavenging. It was a square rock with these weird items on the right and a transparent rock on the left.